Updated Changelog, Finished refactor
