+++ src/BLE_Custom_Characteristic.cpp
@@ -6,6 +6,7 @@
  */
 
 /*
+
 Custom Characteristic for userConfig Variable manipulation via BLE
 
 An example follows to read/write 26.3kph to simulatedSpeed:
@@ -37,25 +38,37 @@ True values are >00. False are 00.
 */
 
 #include <BLE_Common.h>
-#include <Custom_Characteristic.h>
+#include <BLE_Custom_Characteristic.h>
 #include <Constants.h>
+
+void BLE_ss2kCustomCharacteristic::setupService(NimBLEServer *pServer) {
+  pSmartSpin2kService = spinBLEServer.pServer->createService(SMARTSPIN2K_SERVICE_UUID);
+  smartSpin2kCharacteristic =
+      pSmartSpin2kService->createCharacteristic(SMARTSPIN2K_CHARACTERISTIC_UUID, NIMBLE_PROPERTY::WRITE | NIMBLE_PROPERTY::INDICATE | NIMBLE_PROPERTY::NOTIFY);
+  smartSpin2kCharacteristic->setValue(ss2kCustomCharacteristicValue, sizeof(ss2kCustomCharacteristicValue));
+  smartSpin2kCharacteristic->setCallbacks(new ss2kCustomCharacteristicCallbacks());
+  pSmartSpin2kService->start();
+}
+
+void BLE_ss2kCustomCharacteristic::update() {}
 
 void ss2kCustomCharacteristicCallbacks::onWrite(BLECharacteristic *pCharacteristic) {
   std::string rxValue = pCharacteristic->getValue();
-  ss2kCustomCharacteristic::process(rxValue);
+  BLE_ss2kCustomCharacteristic::process(rxValue);
 }
 
-void ss2kCustomCharacteristicCallbacks::onSubscribe(NimBLECharacteristic* pCharacteristic, ble_gap_conn_desc* desc, uint16_t subValue){
-    NimBLEDevice::setMTU(515);
-}
+void ss2kCustomCharacteristicCallbacks::onSubscribe(NimBLECharacteristic *pCharacteristic, ble_gap_conn_desc *desc, uint16_t subValue) { NimBLEDevice::setMTU(515); }
 
-void ss2kCustomCharacteristic::notify(char _item) {
+void BLE_ss2kCustomCharacteristic::notify(char _item) {
   std::string returnValue = {cc_read, _item};
   process(returnValue);
 }
 
-void ss2kCustomCharacteristic::process(std::string rxValue) {
-  // FInd the Characteristic
+void BLE_ss2kCustomCharacteristic::process(std::string rxValue) {
+  // Find the Characteristic
+  if(NimBLEDevice::getServer()->getServiceByUUID(SMARTSPIN2K_SERVICE_UUID) == nullptr){
+    return;
+  }
   NimBLECharacteristic *pCharacteristic = NimBLEDevice::getServer()->getServiceByUUID(SMARTSPIN2K_SERVICE_UUID)->getCharacteristic(SMARTSPIN2K_CHARACTERISTIC_UUID);
   uint8_t *pData                        = reinterpret_cast<uint8_t *>(&rxValue[0]);
   int length                            = rxValue.length();
@@ -480,7 +493,7 @@ void ss2kCustomCharacteristic::process(std::string rxValue) {
       logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "<-reset to defaults");
       if (rxValue[0] == cc_write) {
         ss2k->resetDefaultsFlag = true;
-        returnValue[0] = cc_success;
+        returnValue[0]          = cc_success;
       }
 
       break;
@@ -601,119 +614,119 @@ void ss2kCustomCharacteristic::process(std::string rxValue) {
 }
 
 // iterate through all smartspin user parameters and notify the specific one if changed
-void ss2kCustomCharacteristic::parseNemit() {
+void BLE_ss2kCustomCharacteristic::parseNemit() {
   static userParameters _oldParams;
 
   if (userConfig->getAutoUpdate() != _oldParams.getAutoUpdate()) {
     _oldParams.setAutoUpdate(userConfig->getAutoUpdate());
-    ss2kCustomCharacteristic::notify(BLE_autoUpdate);
+    BLE_ss2kCustomCharacteristic::notify(BLE_autoUpdate);
     return;  // only do one at a time because immediate update isn't super important for these values
   }
 
   if (strcmp(userConfig->getFirmwareUpdateURL(), _oldParams.getFirmwareUpdateURL()) != 0) {
     _oldParams.setFirmwareUpdateURL(userConfig->getFirmwareUpdateURL());
-    ss2kCustomCharacteristic::notify(BLE_firmwareUpdateURL);
+    BLE_ss2kCustomCharacteristic::notify(BLE_firmwareUpdateURL);
     return;
   }
 
   if (strcmp(userConfig->getDeviceName(), _oldParams.getDeviceName()) != 0) {
     _oldParams.setDeviceName(userConfig->getDeviceName());
-    ss2kCustomCharacteristic::notify(BLE_deviceName);
+    BLE_ss2kCustomCharacteristic::notify(BLE_deviceName);
     return;
   }
 
   if (userConfig->getShiftStep() != _oldParams.getShiftStep()) {
     _oldParams.setShiftStep(userConfig->getShiftStep());
-    ss2kCustomCharacteristic::notify(BLE_shiftStep);
+    BLE_ss2kCustomCharacteristic::notify(BLE_shiftStep);
     return;
   }
 
   if (userConfig->getStealthChop() != _oldParams.getStealthChop()) {
     _oldParams.setStealthChop(userConfig->getStealthChop());
-    ss2kCustomCharacteristic::notify(BLE_stealthChop);
+    BLE_ss2kCustomCharacteristic::notify(BLE_stealthChop);
     return;
   }
 
   if (userConfig->getInclineMultiplier() != _oldParams.getInclineMultiplier()) {
     _oldParams.setInclineMultiplier(userConfig->getInclineMultiplier());
-    ss2kCustomCharacteristic::notify(BLE_inclineMultiplier);
+    BLE_ss2kCustomCharacteristic::notify(BLE_inclineMultiplier);
     return;
   }
 
   if (userConfig->getPowerCorrectionFactor() != _oldParams.getPowerCorrectionFactor()) {
     _oldParams.setPowerCorrectionFactor(userConfig->getPowerCorrectionFactor());
-    ss2kCustomCharacteristic::notify(BLE_powerCorrectionFactor);
+    BLE_ss2kCustomCharacteristic::notify(BLE_powerCorrectionFactor);
     return;
   }
 
   if (strcmp(userConfig->getSsid(), _oldParams.getSsid()) != 0) {
     _oldParams.setSsid(userConfig->getSsid());
-    ss2kCustomCharacteristic::notify(BLE_ssid);
+    BLE_ss2kCustomCharacteristic::notify(BLE_ssid);
     return;
   }
 
   if (strcmp(userConfig->getPassword(), _oldParams.getPassword()) != 0) {
     _oldParams.setPassword(userConfig->getPassword());
-    ss2kCustomCharacteristic::notify(BLE_password);
+    BLE_ss2kCustomCharacteristic::notify(BLE_password);
     return;
   }
 
   if (strcmp(userConfig->getConnectedPowerMeter(), _oldParams.getConnectedPowerMeter()) != 0) {
     _oldParams.setConnectedPowerMeter(userConfig->getConnectedPowerMeter());
-    ss2kCustomCharacteristic::notify(BLE_connectedPowerMeter);
+    BLE_ss2kCustomCharacteristic::notify(BLE_connectedPowerMeter);
     return;
   }
 
   if (strcmp(userConfig->getConnectedHeartMonitor(), _oldParams.getConnectedHeartMonitor()) != 0) {
     _oldParams.setConnectedHeartMonitor(userConfig->getConnectedHeartMonitor());
-    ss2kCustomCharacteristic::notify(BLE_connectedHeartMonitor);
+    BLE_ss2kCustomCharacteristic::notify(BLE_connectedHeartMonitor);
     return;
   }
 
   if (userConfig->getStepperPower() != _oldParams.getStepperPower()) {
     _oldParams.setStepperPower(userConfig->getStepperPower());
-    ss2kCustomCharacteristic::notify(BLE_stepperPower);
+    BLE_ss2kCustomCharacteristic::notify(BLE_stepperPower);
     return;
   }
 
   if (userConfig->getStepperSpeed() != _oldParams.getStepperSpeed()) {
     _oldParams.setStepperSpeed(userConfig->getStepperSpeed());
-    ss2kCustomCharacteristic::notify(BLE_stepperSpeed);
+    BLE_ss2kCustomCharacteristic::notify(BLE_stepperSpeed);
     return;
   }
 
   if (userConfig->getERGSensitivity() != _oldParams.getERGSensitivity()) {
     _oldParams.setERGSensitivity(userConfig->getERGSensitivity());
-    ss2kCustomCharacteristic::notify(BLE_ERGSensitivity);
+    BLE_ss2kCustomCharacteristic::notify(BLE_ERGSensitivity);
     return;
   }
 
   if (userConfig->getStepperDir() != _oldParams.getStepperDir()) {
     _oldParams.setStepperDir(userConfig->getStepperDir());
-    ss2kCustomCharacteristic::notify(BLE_shiftDir);
+    BLE_ss2kCustomCharacteristic::notify(BLE_shiftDir);
     return;
   }
 
   if (strcmp(userConfig->getFoundDevices(), _oldParams.getFoundDevices()) != 0) {
     _oldParams.setFoundDevices(userConfig->getFoundDevices());
-    ss2kCustomCharacteristic::notify(BLE_foundDevices);
+    BLE_ss2kCustomCharacteristic::notify(BLE_foundDevices);
     return;
   }
 
   if (userConfig->getMinWatts() != _oldParams.getMinWatts()) {
     _oldParams.setMinWatts(userConfig->getMinWatts());
-    ss2kCustomCharacteristic::notify(BLE_minBrakeWatts);
+    BLE_ss2kCustomCharacteristic::notify(BLE_minBrakeWatts);
     return;
   }
 
   if (userConfig->getMaxWatts() != _oldParams.getMaxWatts()) {
     _oldParams.setMaxWatts(userConfig->getMaxWatts());
-    ss2kCustomCharacteristic::notify(BLE_maxBrakeWatts);
+    BLE_ss2kCustomCharacteristic::notify(BLE_maxBrakeWatts);
     return;
   }
   if (userConfig->getShifterDir() != _oldParams.getShifterDir()) {
     _oldParams.setShifterDir(userConfig->getShifterDir());
-    ss2kCustomCharacteristic::notify(BLE_shiftDir);
+    BLE_ss2kCustomCharacteristic::notify(BLE_shiftDir);
     return;
   }
 }
