+++ src/BLE_Server.cpp
@@ -11,11 +11,12 @@
 #include "BLE_Cycling_Speed_Cadence.h"
 #include "BLE_Cycling_Power_Service.h"
 #include "BLE_Heart_Service.h"
+#include "BLE_Fitness_Machine_Service.h"
+#include "BLE_Custom_Characteristic.h"
 
 #include <ArduinoJson.h>
 #include <Constants.h>
 #include <NimBLEDevice.h>
-#include <Custom_Characteristic.h>
 #include <cmath>
 #include <limits>
 
@@ -27,103 +28,24 @@ static MyCallbacks chrCallbacks;
 BLE_Cycling_Speed_Cadence cyclingSpeedCadenceService;
 BLE_Cycling_Power_Service cyclingPowerService;
 BLE_Heart_Service heartService;
-
-BLEService *pFitnessMachineService;
-BLECharacteristic *fitnessMachineFeature;
-BLECharacteristic *fitnessMachineIndoorBikeData;
-BLECharacteristic *fitnessMachineStatusCharacteristic;
-BLECharacteristic *fitnessMachineControlPoint;
-BLECharacteristic *fitnessMachineResistanceLevelRange;
-BLECharacteristic *fitnessMachinePowerRange;
-BLECharacteristic *fitnessMachineInclinationRange;
-BLECharacteristic *fitnessMachineTrainingStatus;
-
-BLEService *pSmartSpin2kService;
-BLECharacteristic *smartSpin2kCharacteristic;
-std::string FTMSWrite = "";
-
-// Fitness Machine
-uint8_t ftmsIndoorBikeData[11] = {0};
-
-// Resistance, IPower, HeartRate
-uint8_t ftmsResistanceLevelRange[6]      = {0x01, 0x00, 0x64, 0x00, 0x01, 0x00};  // 1:100 increment 1
-uint8_t ftmsPowerRange[6]                = {0x01, 0x00, 0xA0, 0x0F, 0x01, 0x00};  // 1:4000 watts increment 1
-uint8_t ftmsInclinationRange[6]          = {0x38, 0xff, 0xc8, 0x00, 0x01, 0x00};  // -20.0:20.0 increment .1
-uint8_t ftmsTrainingStatus[2]            = {0x08, 0x00};
-uint8_t ss2kCustomCharacteristicValue[3] = {0x00, 0x00, 0x00};
-
-void logCharacteristic(char *buffer, const size_t bufferCapacity, const byte *data, const size_t dataLength, const NimBLEUUID serviceUUID, const NimBLEUUID charUUID,
-                       const char *format, ...) {
-  int bufferLength = ss2k_log_hex_to_buffer(data, dataLength, buffer, 0, bufferCapacity);
-  bufferLength += snprintf(buffer + bufferLength, bufferCapacity - bufferLength, "-> %s | %s | ", serviceUUID.toString().c_str(), charUUID.toString().c_str());
-  va_list args;
-  va_start(args, format);
-  bufferLength += vsnprintf(buffer + bufferLength, bufferCapacity - bufferLength, format, args);
-  va_end(args);
-
-  SS2K_LOG(BLE_SERVER_LOG_TAG, "%s", buffer);
-#ifdef USE_TELEGRAM
-  SEND_TO_TELEGRAM(String(buffer));
-#endif
-}
+BLE_Fitness_Machine_Service fitnessMachineService;
+BLE_ss2kCustomCharacteristic ss2kCustomCharacteristic;
 
 void startBLEServer() {
   // Server Setup
   SS2K_LOG(BLE_SERVER_LOG_TAG, "Starting BLE Server");
   spinBLEServer.pServer = BLEDevice::createServer();
-
-  // Fitness Machine Feature Flags Setup
-  struct FitnessMachineFeature ftmsFeature = {FitnessMachineFeatureFlags::Types::CadenceSupported | FitnessMachineFeatureFlags::Types::HeartRateMeasurementSupported |
-                                                  FitnessMachineFeatureFlags::Types::PowerMeasurementSupported | FitnessMachineFeatureFlags::Types::InclinationSupported |
-                                                  FitnessMachineFeatureFlags::Types::ResistanceLevelSupported,
-                                              FitnessMachineTargetFlags::PowerTargetSettingSupported | FitnessMachineTargetFlags::Types::InclinationTargetSettingSupported |
-                                                  FitnessMachineTargetFlags::Types::ResistanceTargetSettingSupported |
-                                                  FitnessMachineTargetFlags::Types::IndoorBikeSimulationParametersSupported |
-                                                  FitnessMachineTargetFlags::Types::SpinDownControlSupported};
-  // Fitness Machine Indoor Bike Data Flags Setup
-  FitnessMachineIndoorBikeDataFlags::Types ftmsIBDFlags = FitnessMachineIndoorBikeDataFlags::InstantaneousCadencePresent |
-                                                          FitnessMachineIndoorBikeDataFlags::ResistanceLevelPresent | FitnessMachineIndoorBikeDataFlags::InstantaneousPowerPresent |
-                                                          FitnessMachineIndoorBikeDataFlags::HeartRatePresent;
-
-  // Fitness Machine service setup
-  pFitnessMachineService             = spinBLEServer.pServer->createService(FITNESSMACHINESERVICE_UUID);
-  fitnessMachineFeature              = pFitnessMachineService->createCharacteristic(FITNESSMACHINEFEATURE_UUID, NIMBLE_PROPERTY::READ);
-  fitnessMachineControlPoint         = pFitnessMachineService->createCharacteristic(FITNESSMACHINECONTROLPOINT_UUID, NIMBLE_PROPERTY::WRITE | NIMBLE_PROPERTY::INDICATE);
-  fitnessMachineStatusCharacteristic = pFitnessMachineService->createCharacteristic(FITNESSMACHINESTATUS_UUID, NIMBLE_PROPERTY::NOTIFY);
-  fitnessMachineIndoorBikeData       = pFitnessMachineService->createCharacteristic(FITNESSMACHINEINDOORBIKEDATA_UUID, NIMBLE_PROPERTY::READ | NIMBLE_PROPERTY::NOTIFY);
-  fitnessMachineResistanceLevelRange = pFitnessMachineService->createCharacteristic(FITNESSMACHINERESISTANCELEVELRANGE_UUID, NIMBLE_PROPERTY::READ);
-  fitnessMachinePowerRange           = pFitnessMachineService->createCharacteristic(FITNESSMACHINEPOWERRANGE_UUID, NIMBLE_PROPERTY::READ);
-  fitnessMachineInclinationRange     = pFitnessMachineService->createCharacteristic(FITNESSMACHINEINCLINATIONRANGE_UUID, NIMBLE_PROPERTY::READ);
-  fitnessMachineTrainingStatus       = pFitnessMachineService->createCharacteristic(FITNESSMACHINETRAININGSTATUS_UUID, NIMBLE_PROPERTY::NOTIFY);
-
-  pSmartSpin2kService = spinBLEServer.pServer->createService(SMARTSPIN2K_SERVICE_UUID);
-  smartSpin2kCharacteristic =
-      pSmartSpin2kService->createCharacteristic(SMARTSPIN2K_CHARACTERISTIC_UUID, NIMBLE_PROPERTY::WRITE | NIMBLE_PROPERTY::INDICATE | NIMBLE_PROPERTY::NOTIFY);
-
   spinBLEServer.pServer->setCallbacks(new MyServerCallbacks());
 
-  fitnessMachineFeature->setValue(ftmsFeature.bytes, sizeof(ftmsFeature));
-  ftmsIndoorBikeData[0] = static_cast<uint8_t>(ftmsIBDFlags & 0xFF);         // LSB, mask with 0xFF to get the lower 8 bits
-  ftmsIndoorBikeData[1] = static_cast<uint8_t>((ftmsIBDFlags >> 8) & 0xFF);  // MSB, shift right by 8 bits and mask with 0xFF
-  fitnessMachineIndoorBikeData->setValue(ftmsIndoorBikeData, sizeof(ftmsIndoorBikeData));
-  fitnessMachineResistanceLevelRange->setValue(ftmsResistanceLevelRange, sizeof(ftmsResistanceLevelRange));
-  fitnessMachinePowerRange->setValue(ftmsPowerRange, sizeof(ftmsPowerRange));
-  fitnessMachineInclinationRange->setValue(ftmsInclinationRange, sizeof(ftmsInclinationRange));
-  smartSpin2kCharacteristic->setValue(ss2kCustomCharacteristicValue, sizeof(ss2kCustomCharacteristicValue));
-
-  fitnessMachineIndoorBikeData->setCallbacks(&chrCallbacks);
-  fitnessMachineControlPoint->setCallbacks(&chrCallbacks);
-  smartSpin2kCharacteristic->setCallbacks(new ss2kCustomCharacteristicCallbacks());
-
+  // start services
   cyclingSpeedCadenceService.setupService(spinBLEServer.pServer, &chrCallbacks);
   cyclingPowerService.setupService(spinBLEServer.pServer, &chrCallbacks);
   heartService.setupService(spinBLEServer.pServer, &chrCallbacks);
+  fitnessMachineService.setupService(spinBLEServer.pServer, &chrCallbacks);
+  ss2kCustomCharacteristic.setupService(spinBLEServer.pServer);
 
-  pFitnessMachineService->start();
-  pSmartSpin2kService->start();
-
-  // const std::string fitnessData = {0b00000001, 0b00100000, 0b00000000};
   BLEAdvertising *pAdvertising = BLEDevice::getAdvertising();
+  // const std::string fitnessData = {0b00000001, 0b00100000, 0b00000000};
   // pAdvertising->setServiceData(FITNESSMACHINESERVICE_UUID, fitnessData);
 
   pAdvertising->addServiceUUID(FITNESSMACHINESERVICE_UUID);
@@ -141,61 +63,25 @@ void startBLEServer() {
   SS2K_LOG(BLE_SERVER_LOG_TAG, "Bluetooth Characteristic defined!");
 }
 
-bool spinDown() {
-  std::string rxValue = fitnessMachineStatusCharacteristic->getValue();
-  if (rxValue[0] != 0x14) {
-    return false;
-  }
-  uint8_t spinStatus[2] = {0x14, 0x01};
-
-  if (rxValue[1] == 0x01) {
-    // debugDirector("Spin Down Initiated", true);
-    vTaskDelay(1000 / portTICK_RATE_MS);
-    spinStatus[1] = 0x04;  // send Stop Pedaling
-    fitnessMachineStatusCharacteristic->setValue(spinStatus, 2);
-  }
-  if (rxValue[1] == 0x04) {
-    // debugDirector("Stop Pedaling", true);
-    vTaskDelay(1000 / portTICK_RATE_MS);
-    spinStatus[1] = 0x02;  // Success
-    fitnessMachineStatusCharacteristic->setValue(spinStatus, 2);
-  }
-  if (rxValue[1] == 0x02) {
-    // debugDirector("Success", true);
-    spinStatus[0] = 0x00;
-    spinStatus[1] = 0x00;  // Success
-    fitnessMachineStatusCharacteristic->setValue(spinStatus, 2);
-    uint8_t returnValue[3] = {0x00, 0x00, 0x00};
-    fitnessMachineControlPoint->setValue(returnValue, 3);
-    fitnessMachineControlPoint->indicate();
-  }
-
-  fitnessMachineStatusCharacteristic->notify();
-
-  return true;
-}
-
-void updateServerChars() {
-  // Setup the information
-  updateWheelAndCrankRev();
+void SpinBLEServer::update() {
+  // Wheel and crank is used in multiple characteristics. Update first.
+  spinBLEServer.updateWheelAndCrankRev();
   // update the BLE information on the server
-  updateIndoorBikeDataChar();
   heartService.update();
   cyclingPowerService.update();
   cyclingSpeedCadenceService.update();
-  // controlPointIndicate();
+  fitnessMachineService.update();
 }
 
-double calculateSpeed() {
+double SpinBLEServer::calculateSpeed() {
   // Constants for the formula: adjusted for calibration
   const double dragCoefficient   = 1.95;
   const double frontalArea       = 0.9;    // m^2
   const double airDensity        = 1.225;  // kg/m^3
   const double rollingResistance = 0.004;
   const double combinedConstant  = 0.5 * airDensity * dragCoefficient * frontalArea + rollingResistance;
-
-  double power                  = rtConfig->watts.getValue();           // Power in watts
-  double speedInMetersPerSecond = std::cbrt(power / combinedConstant);  // Speed in m/s
+  double power                   = rtConfig->watts.getValue();           // Power in watts
+  double speedInMetersPerSecond  = std::cbrt(power / combinedConstant);  // Speed in m/s
 
   // Convert speed from m/s to km/h
   double speedKmH = speedInMetersPerSecond * 3.6;
@@ -207,13 +93,13 @@ double calculateSpeed() {
   return speedKmH;
 }
 
-void updateWheelAndCrankRev() {
+void SpinBLEServer::updateWheelAndCrankRev() {
   float wheelSize     = 2.127;  // 700cX28 circumference, typical in meters
   float wheelSpeedMps = 0.0;
   if (rtConfig->getSimulatedSpeed() > 5) {
     wheelSpeedMps = rtConfig->getSimulatedSpeed() / 3.6;
   } else {
-    wheelSpeedMps = calculateSpeed() / 3.6;  // covert km/h to m/s
+    wheelSpeedMps = this->calculateSpeed() / 3.6;  // covert km/h to m/s
   }
 
   // Calculate wheel revolutions per minute
@@ -230,47 +116,6 @@ void updateWheelAndCrankRev() {
     spinBLEClient.cscCumulativeCrankRev++;
     spinBLEClient.cscLastCrankEvtTime += crankRevPeriod;
   }
-}
-
-void updateIndoorBikeDataChar() {
-  if (!spinBLEServer.clientSubscribed.IndoorBikeData) {
-    return;
-  }
-  float cadRaw      = rtConfig->cad.getValue();
-  int cad           = static_cast<int>(cadRaw * 2);
-  int watts         = rtConfig->watts.getValue();
-  int hr            = rtConfig->hr.getValue();
-  int res           = rtConfig->resistance.getValue();
-  int speedFtmsUnit = 0;
-  if (rtConfig->getSimulatedSpeed() > 5) {
-    speedFtmsUnit = rtConfig->getSimulatedSpeed() * 100;
-  } else {
-    speedFtmsUnit = calculateSpeed() * 100;
-  }
-
-  ftmsIndoorBikeData[2] = (uint8_t)(speedFtmsUnit & 0xff);
-  ftmsIndoorBikeData[3] = (uint8_t)(speedFtmsUnit >> 8);
-
-  ftmsIndoorBikeData[4] = (uint8_t)(cad & 0xff);
-  ftmsIndoorBikeData[5] = (uint8_t)(cad >> 8);
-
-  ftmsIndoorBikeData[6] = (uint8_t)(res & 0xff);
-  ftmsIndoorBikeData[7] = (uint8_t)(res >> 8);
-
-  ftmsIndoorBikeData[8] = (uint8_t)(watts & 0xff);
-  ftmsIndoorBikeData[9] = (uint8_t)(watts >> 8);
-
-  ftmsIndoorBikeData[10] = (uint8_t)hr;
-
-  fitnessMachineIndoorBikeData->setValue(ftmsIndoorBikeData, 11);
-  fitnessMachineIndoorBikeData->notify();
-
-  const int kLogBufCapacity = 200;  // Data(30), Sep(data/2), Arrow(3), CharId(37), Sep(3), CharId(37), Sep(3), Name(10), Prefix(2), HR(7), SEP(1), CD(10), SEP(1), PW(8),
-                                    // SEP(1), SD(7), Suffix(2), Nul(1), rounded up
-  char logBuf[kLogBufCapacity];
-  const size_t ftmsIndoorBikeDataLength = sizeof(ftmsIndoorBikeData) / sizeof(ftmsIndoorBikeData[0]);
-  logCharacteristic(logBuf, kLogBufCapacity, ftmsIndoorBikeData, ftmsIndoorBikeDataLength, FITNESSMACHINESERVICE_UUID, fitnessMachineIndoorBikeData->getUUID(),
-                    "FTMS(IBD)[ HR(%d) CD(%.2f) PW(%d) SD(%.2f) ]", hr % 1000, fmodf(cadRaw, 1000.0), watts % 10000, fmodf(speedFtmsUnit / 100, 1000.0));
 }
 
 // Creating Server Connection Callbacks
@@ -300,7 +145,15 @@ bool MyServerCallbacks::onConnParamsUpdateRequest(NimBLEClient *pClient, const b
   return true;
 };
 
-void MyCallbacks::onWrite(BLECharacteristic *pCharacteristic) { FTMSWrite = pCharacteristic->getValue(); }
+// END SERVER CALLBACKS
+
+void MyCallbacks::onWrite(BLECharacteristic *pCharacteristic) {
+  if (pCharacteristic->getUUID() == FITNESSMACHINECONTROLPOINT_UUID) {
+    spinBLEServer.writeCache.push(pCharacteristic->getValue());
+  } else {
+    SS2K_LOG(BLE_SERVER_LOG_TAG, "Write to %s is not supported", pCharacteristic->getUUID().toString());
+  }
+}
 
 void MyCallbacks::onSubscribe(NimBLECharacteristic *pCharacteristic, ble_gap_conn_desc *desc, uint16_t subValue) {
   String str       = "Client ID: ";
@@ -337,217 +190,6 @@ void SpinBLEServer::setClientSubscribed(NimBLEUUID pUUID, bool subscribe) {
     spinBLEServer.clientSubscribed.CyclingSpeedCadence = subscribe;
   }
 }
-
-// The things that happen when we receive a FitnessMachineControlPointProcedure from a Client.
-void processFTMSWrite() {
-  if (FTMSWrite == "") {
-    return;
-  }
-  BLECharacteristic *pCharacteristic = NimBLEDevice::getServer()->getServiceByUUID(FITNESSMACHINESERVICE_UUID)->getCharacteristic(FITNESSMACHINECONTROLPOINT_UUID);
-
-  std::string rxValue = FTMSWrite;
-  std::vector<uint8_t> ftmsStatus;
-  if (rxValue.length() >= 1) {
-    uint8_t *pData = reinterpret_cast<uint8_t *>(&rxValue[0]);
-    int length     = rxValue.length();
-
-    const int kLogBufCapacity = (rxValue.length() * 2) + 60;  // largest comment is 48 VV
-    char logBuf[kLogBufCapacity];
-    int logBufLength       = ss2k_log_hex_to_buffer(pData, length, logBuf, 0, kLogBufCapacity);
-    int port               = 0;
-    uint8_t returnValue[3] = {FitnessMachineControlPointProcedure::ResponseCode, (uint8_t)rxValue[0], FitnessMachineControlPointResultCode::OpCodeNotSupported};
-
-    ftmsStatus = {FitnessMachineStatus::ReservedForFutureUse};
-
-    switch ((uint8_t)rxValue[0]) {
-      case FitnessMachineControlPointProcedure::RequestControl:
-        returnValue[2] = FitnessMachineControlPointResultCode::Success;  // 0x01;
-        pCharacteristic->setValue(returnValue, 3);
-        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Control Request");
-        ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Idle;  // 0x01;
-        fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-        ftmsStatus = {FitnessMachineStatus::StartedOrResumedByUser};
-        break;
-
-      case FitnessMachineControlPointProcedure::Reset: {
-        returnValue[2] = FitnessMachineControlPointResultCode::Success;  // 0x01;
-        pCharacteristic->setValue(returnValue, 3);
-
-        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Reset");
-        ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Idle;  // 0x01;
-        ftmsStatus            = {FitnessMachineStatus::Reset};
-        fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-      } break;
-
-      case FitnessMachineControlPointProcedure::SetTargetInclination: {
-        rtConfig->setFTMSMode((uint8_t)rxValue[0]);
-        returnValue[2] = FitnessMachineControlPointResultCode::Success;  // 0x01;
-        pCharacteristic->setValue(returnValue, 3);
-
-        port = (rxValue[2] << 8) + rxValue[1];
-        port *= 10;
-
-        rtConfig->setTargetIncline(port);
-
-        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Incline Mode: %2f", rtConfig->getTargetIncline() / 100);
-
-        ftmsStatus            = {FitnessMachineStatus::TargetInclineChanged, (uint8_t)rxValue[1], (uint8_t)rxValue[2]};
-        ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Other;  // 0x00;
-        fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-      } break;
-
-      case FitnessMachineControlPointProcedure::SetTargetResistanceLevel: {
-        rtConfig->setFTMSMode((uint8_t)rxValue[0]);
-        returnValue[2]        = FitnessMachineControlPointResultCode::Success;  // 0x01;
-        ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Other;            // 0x00;
-        fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-        if ((int)rxValue[1] >= rtConfig->getMinResistance() && (int)rxValue[1] <= rtConfig->getMaxResistance()) {
-          rtConfig->resistance.setTarget((int)rxValue[1]);
-          returnValue[2] = FitnessMachineControlPointResultCode::Success;
-
-          logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Resistance Mode: %d", rtConfig->resistance.getTarget());
-        } else if ((int)rxValue[1] > rtConfig->getMinResistance()) {
-          rtConfig->resistance.setTarget(rtConfig->getMaxResistance());
-          returnValue[2] = FitnessMachineControlPointResultCode::InvalidParameter;
-          logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Resistance Request %d beyond limits", (int)rxValue[1]);
-        } else {
-          rtConfig->resistance.setTarget(rtConfig->getMinResistance());
-          returnValue[2] = FitnessMachineControlPointResultCode::InvalidParameter;
-          logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Resistance Request %d beyond limits", (int)rxValue[1]);
-        }
-        ftmsStatus = {FitnessMachineStatus::TargetResistanceLevelChanged, (uint8_t)(rtConfig->resistance.getTarget() % 256)};
-        rtConfig->resistance.setTarget(rtConfig->resistance.getTarget());
-        pCharacteristic->setValue(returnValue, 3);
-      } break;
-
-      case FitnessMachineControlPointProcedure::SetTargetPower: {
-        rtConfig->setFTMSMode((uint8_t)rxValue[0]);
-        if (spinBLEClient.connectedPM || rtConfig->watts.getSimulate()) {
-          returnValue[2] = FitnessMachineControlPointResultCode::Success;  // 0x01;
-
-          rtConfig->watts.setTarget(bytes_to_u16(rxValue[2], rxValue[1]));
-          logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> ERG Mode Target: %d Current: %d Incline: %2f", rtConfig->watts.getTarget(),
-                                   rtConfig->watts.getValue(), rtConfig->getTargetIncline() / 100);
-
-          ftmsStatus            = {FitnessMachineStatus::TargetPowerChanged, (uint8_t)rxValue[1], (uint8_t)rxValue[2]};
-          ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::WattControl;  // 0x0C;
-          fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-          // Adjust set point for powerCorrectionFactor and send to FTMS server (if connected)
-          int adjustedTarget         = rtConfig->watts.getTarget() / userConfig->getPowerCorrectionFactor();
-          const uint8_t translated[] = {FitnessMachineControlPointProcedure::SetTargetPower, (uint8_t)(adjustedTarget % 256), (uint8_t)(adjustedTarget / 256)};
-          spinBLEClient.FTMSControlPointWrite(translated, 3);
-        } else {
-          returnValue[2] = FitnessMachineControlPointResultCode::OpCodeNotSupported;  // 0x02; no power meter connected, so no ERG
-          logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> ERG Mode: No Power Meter Connected");
-        }
-        pCharacteristic->setValue(returnValue, 3);
-      } break;
-
-      case FitnessMachineControlPointProcedure::StartOrResume: {
-        returnValue[2] = FitnessMachineControlPointResultCode::Success;  // 0x01;
-        pCharacteristic->setValue(returnValue, 3);
-
-        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Start Training");
-        ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Other;  // 0x00;
-        fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-        ftmsStatus = {FitnessMachineStatus::StartedOrResumedByUser};
-      } break;
-
-      case FitnessMachineControlPointProcedure::StopOrPause: {
-        returnValue[2] = FitnessMachineControlPointResultCode::Success;  // 0x01;
-        pCharacteristic->setValue(returnValue, 3);
-        // rxValue[1] == 1 -> Stop, 2 -> Pause
-        // TODO: Move stepper to Min Position
-        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Stop Training");
-
-        ftmsStatus            = {FitnessMachineStatus::StoppedOrPausedByUser};
-        ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Other;  // 0x00;
-        fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-
-      } break;
-
-      case FitnessMachineControlPointProcedure::SetIndoorBikeSimulationParameters: {  // sim mode
-        rtConfig->setFTMSMode((uint8_t)rxValue[0]);
-        returnValue[2] = FitnessMachineControlPointResultCode::Success;  // 0x01;
-        pCharacteristic->setValue(returnValue, 3);
-
-        signed char buf[2];
-        // int16_t windSpeed        = (rxValue[2] << 8) + rxValue[1];
-        buf[0] = rxValue[3];  // (Least significant byte)
-        buf[1] = rxValue[4];  // (Most significant byte)
-        // int8_t rollingResistance = rxValue[5];
-        // int8_t windResistance    = rxValue[6];
-        port = bytes_to_u16(buf[1], buf[0]);
-        rtConfig->setTargetIncline(port);
-        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Sim Mode Incline %2f", rtConfig->getTargetIncline() / 100);
-
-        ftmsStatus = {FitnessMachineStatus::IndoorBikeSimulationParametersChanged,
-                      (uint8_t)rxValue[1],
-                      (uint8_t)rxValue[2],
-                      (uint8_t)rxValue[3],
-                      (uint8_t)rxValue[4],
-                      (uint8_t)rxValue[5],
-                      (uint8_t)rxValue[6]};
-
-        ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Other;  // 0x00;
-        fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-        spinBLEClient.FTMSControlPointWrite(pData, length);
-      } break;
-
-      case FitnessMachineControlPointProcedure::SpinDownControl: {
-        rtConfig->setFTMSMode((uint8_t)rxValue[0]);
-        uint8_t controlPoint[6] = {FitnessMachineControlPointProcedure::ResponseCode, 0x01, 0x24, 0x03, 0x96, 0x0e};  // send low and high speed targets
-        pCharacteristic->setValue(controlPoint, 6);
-        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Spin Down Requested");
-
-        ftmsStatus = {FitnessMachineStatus::SpinDownStatus, 0x01};  // send low and high speed targets
-
-        ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Other;  // 0x00;
-        fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-      } break;
-
-      case FitnessMachineControlPointProcedure::SetTargetedCadence: {
-        rtConfig->setFTMSMode((uint8_t)rxValue[0]);
-        returnValue[2] = FitnessMachineControlPointResultCode::Success;  // 0x01;
-        pCharacteristic->setValue(returnValue, 3);
-
-        int targetCadence = bytes_to_u16(rxValue[2], rxValue[1]);
-        // rtConfig->setTargetCadence(targetCadence);
-        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Target Cadence: %d ", targetCadence);
-
-        ftmsStatus = {FitnessMachineStatus::TargetedCadenceChanged, (uint8_t)rxValue[1], (uint8_t)rxValue[2]};
-
-        ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Other;  // 0x00;
-        fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-      } break;
-
-      default: {
-        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Unsupported FTMS Request");
-        pCharacteristic->setValue(returnValue, 3);
-      }
-    }
-    SS2K_LOG(FMTS_SERVER_LOG_TAG, "%s", logBuf);
-  } else {
-    SS2K_LOG(FMTS_SERVER_LOG_TAG, "App wrote nothing ");
-    SS2K_LOG(FMTS_SERVER_LOG_TAG, "assuming it's a Control request");
-
-    uint8_t controlPoint[3] = {FitnessMachineControlPointProcedure::ResponseCode, 0x00, FitnessMachineControlPointResultCode::Success};
-    pCharacteristic->setValue(controlPoint, 3);
-    ftmsStatus            = {FitnessMachineStatus::StartedOrResumedByUser};
-    ftmsTrainingStatus[1] = FitnessMachineTrainingStatus::Other;  // 0x00;
-    fitnessMachineTrainingStatus->setValue(ftmsTrainingStatus, 2);
-    fitnessMachineTrainingStatus->notify(false);
-  }
-  for (int i = 0; i < ftmsStatus.size(); i++) {
-  }
-  fitnessMachineStatusCharacteristic->setValue(ftmsStatus.data(), ftmsStatus.size());
-  pCharacteristic->indicate();
-  fitnessMachineTrainingStatus->notify(false);
-  fitnessMachineStatusCharacteristic->notify(false);
-  FTMSWrite = "";
-}
-
-void controlPointIndicate() { fitnessMachineControlPoint->indicate(); }
 
 // Return number of clients connected to our server.
 int connectedClientCount() {
@@ -589,4 +231,19 @@ void calculateInstPwrFromHR() {
 #endif  // DEBUG_HR_TO_PWR
 
   SS2K_LOG(BLE_SERVER_LOG_TAG, "Power From HR: %d", avgP);
+}
+
+void logCharacteristic(char *buffer, const size_t bufferCapacity, const byte *data, const size_t dataLength, const NimBLEUUID serviceUUID, const NimBLEUUID charUUID,
+                       const char *format, ...) {
+  int bufferLength = ss2k_log_hex_to_buffer(data, dataLength, buffer, 0, bufferCapacity);
+  bufferLength += snprintf(buffer + bufferLength, bufferCapacity - bufferLength, "-> %s | %s | ", serviceUUID.toString().c_str(), charUUID.toString().c_str());
+  va_list args;
+  va_start(args, format);
+  bufferLength += vsnprintf(buffer + bufferLength, bufferCapacity - bufferLength, format, args);
+  va_end(args);
+
+  SS2K_LOG(BLE_SERVER_LOG_TAG, "%s", buffer);
+#ifdef USE_TELEGRAM
+  SEND_TO_TELEGRAM(String(buffer));
+#endif
 }
